feat(routes): create account & profiles pages
